fix(client): passes FusionAuth client by value.
This fixes an issue where passing the FusionAuth client by reference could lead to tenant ids being set for multiple different requests.

It also alleviates an issue where under concurrency the tenant id from a specific request could lead to being set as the default tenant id for later requests, if a request modifying the client's tenant id was in-flight waiting on a response, while another request is issued.