fix: login records information being incorrectly read from admin logs fields
When using terrraform to make changes to the login records section of the system configuration settings a terraform apply indicates that changes are always required. This is because the when reading the state of the login records settings from Fusion Auth there is a bug in the code that is causing the loging records settings to be read from the admin logs settings fields. I believe these changes should correct this.